fix: when deleting a transaction, an incorrect amount was returned

+++ app/src/main/java/com/yaromchikv/moneymanager/feature/domain/usecase/TransactionUseCases.kt
@@ -71,8 +71,13 @@ class AddTransaction(
     private val accountsRepository: AccountsRepository
 ) {
     suspend operator fun invoke(transaction: Transaction) {
-        transactionsRepository.insertTransaction(transaction)
-        accountsRepository.updateAccountAmountById(transaction.accountId, -1 * transaction.amount)
+        val account = accountsRepository.getAccountById(transaction.accountId)
+        if (account != null) {
+            val amount = transaction.amount + account.amount
+            accountsRepository.updateAccountAmount(transaction.accountId, amount)
+
+            transactionsRepository.insertTransaction(transaction)
+        }
     }
 }
 
@@ -81,7 +86,12 @@ class DeleteTransactionById(
     private val accountsRepository: AccountsRepository
 ) {
     suspend operator fun invoke(transaction: TransactionView) {
-        repository.deleteTransactionById(transaction.id)
-        accountsRepository.updateAccountAmountById(transaction.accountId, transaction.amount)
+        val account = accountsRepository.getAccountById(transaction.accountId)
+        if (account != null) {
+            val amount = account.amount - transaction.amount
+            accountsRepository.updateAccountAmount(transaction.accountId, amount)
+
+            repository.deleteTransactionById(transaction.id)
+        }
     }
 }
